store: Rename 'mapping_sql' module to 'relational_queries'
